% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{add_all_samples}
\alias{add_all_samples}
\title{Add multiple samples to a CHESS S3 object.}
\usage{
add_all_samples(s, sample_list, vaf_noise = 0, genome_size = 3.9e+09, ...)
}
\arguments{
\item{s}{CHESS S3 object.}

\item{sample_list}{A list of samples. Each element must contain a list with arguments that can be passed to the add_sample function (see ?add_sample for details). Try 'random_cells' for testing.}
}
\value{
CHESS S3 object with all the samples added.
}
\description{
Note: Old samples will be removed. Mutations that are detected in a other sample
will also be keeped in all other samples.
}
\examples{
add_all_samples(new_simulation(20, seed=1), "random_cells") # a small 20x20x1 simulation

}
