% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{add_cell_mutation_burden_CHESS_S3}
\alias{add_cell_mutation_burden_CHESS_S3}
\title{Adds each cells mutation burden to a CHESS S3 object.}
\usage{
add_cell_mutation_burden_CHESS_S3(x)
}
\arguments{
\item{x}{A CHESS_S3 object.}
}
\value{
The CHESS_S3 object with the added data.
}
\description{
Adds each cells mutation burden to a CHESS S3 object.
}
\examples{
add_cell_mutation_burden_CHESS_S3(new_simulation(x=20, y=20)) # 20x20 simulation
}
