% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{add_center_position}
\alias{add_center_position}
\title{Estimates the center position of the simulations}
\usage{
add_center_position(x, n = 1000)
}
\arguments{
\item{x}{A CHESS_S3 object.}
}
\value{
The CHESS_S3 object with the center added
}
\description{
Estimates the center position of the simulations
}
\examples{
add_center_position(new_simulation(x=20, y=20)) # 20x20 simulation
}
