% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{add_local_entropy}
\alias{add_local_entropy}
\title{Adds local bulk sample entropy to object}
\usage{
add_local_entropy(s, radius = c(5, 5, 0), min_vaf = 0.05)
}
\arguments{
\item{s}{}

\item{radius}{}

\item{min_vaf}{}
}
\value{

}
\description{
Adds local bulk sample entropy to object
}
\examples{
add_local_entropy(new_simulation(20), radius=c(2,2,0))
}
