% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{add_random_types}
\alias{add_random_types}
\title{Adds cell type data for a subset of cells to a CHESS S3 object.}
\usage{
add_random_types(x, n = 1000)
}
\arguments{
\item{x}{A CHESS_S3 object.}
}
\value{
The CHESS_S3 object with the added data.
}
\description{
Adds cell type data for a subset of cells to a CHESS S3 object.
}
\examples{
add_random_types(new_simulation(x=20, y=20)) # 20x20 simulation
}
