% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{add_sample}
\alias{add_sample}
\alias{addSample}
\title{Adds a sample to the simulation object}
\usage{
add_sample(
  x,
  center,
  diameter = rep(1, length(center)),
  depth = 100,
  depth_model = 1,
  min_reads = 2,
  min_vaf = 0.05,
  seed = 1,
  purity = 1,
  ...
)

addSample(...)
}
\arguments{
\item{x}{CHESS S3 object.}

\item{center}{Center of sample position (1-based index).}

\item{diameter}{Diameter of sample position(1-based index).}

\item{depth}{Average sequencing depth (default 100x).}

\item{depth_model}{Sequencing model.}

\item{min_reads}{Minimum number of reads to keep mutation (default: 1).}

\item{min_vaf}{Minimum VAF to keep mutation (default 0.05)..}

\item{seed}{Seed. Note: A value of 0 indicates to not set a seed for the sample. This will use the state of the random number generator at the time.}

\item{purity}{Sample purity (default: 1.0, pure sample).}

\item{...}{}
}
\value{
CHESS S3 object with the added sample.
}
\description{
Adds a sample to the simulation object

Add multiple samples to a CHESS S3 object
}
\section{Functions}{
\itemize{
\item \code{addSample}: Add multiple samples to a CHESS S3 object
}}

\examples{
add_sample(new_simulation(x=20, y=20, seed=1), center=c(10,10)) # 20x20 simulation

}
