% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_abc_methods.R
\name{calc_ess}
\alias{calc_ess}
\title{Calculate  effective sample size (ESS)}
\usage{
calc_ess(weight)
}
\arguments{
\item{weight}{A numeric vector of particle weights. This will be normalized.}
}
\value{
The effective sample size of the current particle population.
}
\description{
Calculate  effective sample size (ESS)
}
