% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_abc_methods.R
\name{calc_w}
\alias{calc_w}
\title{Calculate particle weights}
\usage{
calc_w(d, epsilon)
}
\arguments{
\item{d}{A nmatrix (2D) or array (3D) of distance values.}

\item{epsilon}{The distance threshold of accepted particles.}
}
\value{
Particle weights.
}
\description{
Calculate particle weights
}
