% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_abc_methods.R
\name{d_matrix_from_particles_pb}
\alias{d_matrix_from_particles_pb}
\title{Calculate distance matrix of a particle set}
\usage{
d_matrix_from_particles_pb(
  p,
  y,
  rho,
  multi_summary = NULL,
  use_precalc_d = TRUE,
  simplify = TRUE,
  n_cores = 1
)
}
\arguments{
\item{p}{The particle set.}

\item{y}{The target tree.}

\item{rho}{The distance function. Must return a proper distance.}

\item{multi_summary}{(optional) Function to simplified multiple observations (default: none).}

\item{use_precalc_d}{(optional) Logical indicating if pre-calculated distance should be trusted (default: true).}

\item{simplify}{(optional) Logical indicating if the distance array should be simplified to a matrix (default: true).}
}
\value{
A 2d matrix (particles, realizations * multi_summary(observations)) or a 3d array (particles, realizations, observations) if "simplify" is FALSE and "multi_summary" is NULL. Missing values (i.e. NA) will be set to Inf!
}
\description{
Calculate distance matrix of a particle set
}
