% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-getters.R
\name{get_global_vaf_sample}
\alias{get_global_vaf_sample}
\title{Gets a matrix of VAF from samples in the CHESS S3 object.}
\usage{
get_global_vaf_sample(x, ...)
}
\arguments{
\item{x}{CHESS S3 object.}

\item{...}{(Optional) addtional arguments passed to the add_sample method (see ?add_sample).}
}
\value{
Returns data.frame with mutation information.
}
\description{
Gets a matrix of VAF from samples in the CHESS S3 object.
}
\examples{
get_global_vaf_sample(new_simulation(20, seed=1), depth=50) # a small 20x20x1 simulation
}
