% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{get_tree}
\alias{get_tree}
\title{Gets tree object from a CHESS S3 simulation object}
\usage{
get_tree(
  x,
  sample_tree = TRUE,
  add_lowdepth = TRUE,
  lp_depth_cutoff = 10,
  tree_manipulator = NULL,
  ...
)
}
\arguments{
\item{x}{The simulation (a CHESS S3 object).}

\item{sample_tree}{(optional) logical flag indicating if the underlying tree should be returned (default: true).}

\item{add_lowdepth}{(optional) logical flag indicating if the low depth samples should be included (default: true).}

\item{lp_depth_cutoff}{(optional) numeric value indicating depth below which a samples is considered to be a low depth sample  (default: 10).}
}
\value{
A phylogenetic tree object (class phylo).
}
\description{
Gets tree object from a CHESS S3 simulation object
}
\examples{
sim = add_all_samples(new_simulation(20, seed=1), "random_cells") # a small 20x20x1 simulation
get_tree(sim, sample_tree=TRUE) # the latent tree
get_tree(sim, sample_tree=FALSE) # the MPR tree from the sequencing data
}
