% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-setters.R
\name{mutation_id_to_state}
\alias{mutation_id_to_state}
\title{Internal function to map mutations to a state vector}
\usage{
mutation_id_to_state(id, weight, seed = 123)
}
\arguments{
\item{id}{Mutation id}

\item{weight}{Vector of number weights}

\item{seed}{optional seed to use for random permutation of mapping.}
}
\value{
vector with length equal to id with state index
}
\description{
Internal function to map mutations to a state vector
}
\examples{
table(mutation_id_to_state(as.character(1:1000), c(4, 1, 5)))
}
