% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-general_methods.R
\name{new_simulation}
\alias{new_simulation}
\title{Constructor for a CHESS simulation S3 object (CHESS_S3).}
\usage{
new_simulation(
  x,
  y = x,
  z = 1,
  param.matrix = NULL,
  seed = NULL,
  verbose = FALSE,
  clonal_mutations = 0,
  snapshot_times = NULL,
  explore_locally = FALSE,
  record_history = TRUE,
  return_generation_tree = FALSE,
  max_popsize = 0,
  n_center_estimate = 0,
  ...
)
}
\arguments{
\item{x}{Size of dimension x. Soft boundary if set to 1.}

\item{y}{Optional size of second dimension (default: same as x).}

\item{z}{Optional size of third dimension (default: 1)..}

\item{param.matrix}{}

\item{seed}{}

\item{verbose}{}

\item{clonal_mutations}{}

\item{snapshot_times}{}

\item{explore_locally}{}

\item{record_history}{}

\item{...}{}
}
\value{

}
\description{
Constructor for a CHESS simulation S3 object (CHESS_S3).
}
\examples{
new_simulation(x=20, y=20) # 20x20 simulation
}
