% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-getters.R
\name{phylo_from_samples}
\alias{phylo_from_samples}
\title{Get phylogenetic data from samples in a CHESS S3 object}
\usage{
phylo_from_samples(
  x,
  all_trees = FALSE,
  add_lowdepth = FALSE,
  min_confidence = 0,
  ...
)
}
\arguments{
\item{x}{A CHESS_S3 object. Must include samples.}

\item{all_trees}{(optional) flag indicating whether all maximum parsimony trees should be returned (default: false).}

\item{add_lowdepth}{(optional) flag indicating whether low coverage samples should be added through a ML method (default: false).}

\item{min_confidence}{(optional) numeric value indicating minimal confidence for the ML method (default: 0.99).}

\item{...}{(optional) arguments passed to the phy_data_from_samples or add_lowpass_sampled function (see ?phy_data_from_samples and ?add_lowpass_sampled for details).}
}
\value{
A phylogentic tree.
}
\description{
Get phylogenetic data from samples in a CHESS S3 object
}
