% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{plot.CHESS_S3}
\alias{plot.CHESS_S3}
\title{Wrapper function arranging multiple plots for a CHESS S3 simulation.}
\usage{
\method{plot}{CHESS_S3}(
  x,
  target_tree = NULL,
  sample_tree = FALSE,
  global_vaf_plot = TRUE,
  run_mobster = FALSE,
  add_history_plot = TRUE,
  plot_tree_function = CHESS::plot_tree,
  modify_tree_function = NULL,
  bulk_depth = 100,
  min_vaf_bulk = 0.05,
  bins_vaf = 100,
  n_clonal = NULL,
  label_drivers = TRUE,
  n_muts_mobster = 20000,
  ...
)
}
\arguments{
\item{x}{A CHESS_S3 object.}

\item{target_tree}{(optional) tree object of class phylo to plot along the data (default: NULL).}

\item{sample_tree}{(optional) flag indicating if simulated sequencing data should be used for the tree reconstruction (default: false).}

\item{global_vaf_plot}{(optional) flag indicating if a global VAF plot should be added (default: true).}

\item{run_mobster}{(optional) flag indicating if mobster should be run on the global vaf data (default: false).}

\item{plot_tree_function}{(optional) alternative function for the tree plotting (default: NULL).}

\item{...}{}
}
\value{
A ggplot object with multiple plots arranged in a grid.
}
\description{
Wrapper function arranging multiple plots for a CHESS S3 simulation.
}
\examples{
plot.CHESS_S3(add_all_samples(new_simulation(20, seed=1), "random_cells")) # a small 20x20x1 simulation, with a tree
plot.CHESS_S3(add_all_samples(new_simulation(20, seed=1), "random_cells"), sample_tree=TRUE) # a small 20x20x1 simulation, with a tree
plot.CHESS_S3(add_all_samples(new_simulation(20, seed=1), "random_cells"), target_tree=ape::rtree(5)) # a small 20x20x1 simulation, with a tree
plot.CHESS_S3(new_simulation(20, seed=1)) # a small 20x20x1 simulation
plot.CHESS_S3(new_simulation(20, seed=1), global_vaf_plot=FALSE) # a small 20x20x1 simulation, don't plot the VAF of mutations
plot.CHESS_S3(new_simulation(20, seed=1), run_mobster=TRUE) # a small 20x20x1 simulation, don't plot the VAF of mutations
}
