% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{plot_history_CHESS_S3}
\alias{plot_history_CHESS_S3}
\title{Plot population sizes over time for a CHESS_S3 simulation.}
\usage{
plot_history_CHESS_S3(x)
}
\arguments{
\item{x}{A CHESS_S3 object.}
}
\value{
A ggplot object.
}
\description{
Plot population sizes over time for a CHESS_S3 simulation.
}
\examples{
plot_history_CHESS_S3(new_simulation(20)) # a small 20x20x1 simulation
}
