% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{plot_space_CHESS_S3}
\alias{plot_space_CHESS_S3}
\title{Plot space of a CHESS_S3 simulation.}
\usage{
plot_space_CHESS_S3(
  x,
  alpha_burden = TRUE,
  mm_per_dot = 0.1,
  labeller_function = NULL,
  rotate_by = 0,
  mirror = FALSE,
  add_transformation_locations = TRUE,
  recenter = FALSE,
  add_polygons = FALSE,
  ...
)
}
\arguments{
\item{x}{A CHESS_S3 object.}

\item{alpha_burden}{(optional) flag indicating if the alpha value should indicate the mutation burden of cells (default: true).}

\item{mm_per_dot}{(optional) value that scales the grid points to a pseudo mm value (default: 0.1).}

\item{labeller_function}{(optional) function that modifies the tip.labels (default: NULL).}

\item{rotate_by}{(optional) numeric value indicating by how many radians the space should be rotated (default: 0).}

\item{mirror}{(optional) flag indicating if the space should be mirrored along the x axis (default: false).}

\item{add_transformation_locations}{}

\item{...}{}
}
\value{
A ggplot object.
}
\description{
Plot space of a CHESS_S3 simulation.
}
\examples{
sim = new_simulation(20, seed=1) # a small 20x20x1 simulation.
sim_plus_sample = add_all_samples(sim, sample_list="random_cells")
plot_space_CHESS_S3(sim) 
plot_space_CHESS_S3(sim, rotate_by=0.5*pi) # as above, but rotated 90° to the left.
plot_space_CHESS_S3(sim, mirror=TRUE) # as above, but mirrored along the x-axis.
plot_space_CHESS_S3(sim_plus_sample) # plot with samples
labeller_function = function(x) LETTERS[seq_along(x)]
plot_space_CHESS_S3(sim_plus_sample, labeller_function=labeller_function) # plot with modified labels
}
