% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{plot_tree}
\alias{plot_tree}
\title{A internal function used for the plotting of trees}
\usage{
plot_tree(tree, HI = NULL)
}
\arguments{
\item{tree}{object of class phylo to plot.}

\item{HI}{(optional) HI of the tree (unused).}
}
\value{
ggplot object
}
\description{
A internal function used for the plotting of trees
}
\examples{
plot_tree(ape::rtree(5))
}
