% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{plot_vaf_sample}
\alias{plot_vaf_sample}
\title{Internal function for plotting of VAF data}
\usage{
plot_vaf_sample(data, n_bins = 100, show_cluster_means = FALSE)
}
\arguments{
\item{data}{data.frame with columns 'vaf' and 'clone'.}
}
\value{
ggplot object.
}
\description{
Internal function for plotting of VAF data
}
\keyword{internal}
