% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{rotate_data}
\alias{rotate_data}
\title{Internal function to rotate grid coordinates in a data.frame}
\usage{
rotate_data(d, angle, center, mirror = FALSE)
}
\arguments{
\item{d}{A data.frame}

\item{angle}{Angle in radians by which to rotate the data.}

\item{center}{Center of the coordinate system around which to rotate.}

\item{mirror}{(Optional) flag indicating if data should be mirrored around the x-axis.}
}
\value{
Data.frame with the rotated coordinates
}
\description{
Internal function to rotate grid coordinates in a data.frame
}
\keyword{internal}
