% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_s3_object-plotting_methods.R
\name{rotate_sim_coords}
\alias{rotate_sim_coords}
\title{Internal function to rotate all coordinat data in a CHESS S3 object}
\usage{
rotate_sim_coords(x, phi = NULL, mirror = FALSE)
}
\arguments{
\item{x}{CHESS S3 object.}

\item{phi}{(Optional) angle in radians by which to rotate the data (default: rotation angle of sample locations).}

\item{mirror}{(Optional) flag indicating if data should be mirrored around the x-axis (default: false).}
}
\value{
CHESS S3 object with modified coordinates.
}
\description{
Internal function to rotate all coordinat data in a CHESS S3 object
}
\keyword{internal}
