% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_abc_methods.R
\name{smc_abc_tree}
\alias{smc_abc_tree}
\title{SMC-ABC on a phylogenetic tree}
\usage{
smc_abc_tree(
  const_params,
  p_gen,
  s_gen,
  target_y,
  rho,
  N = 100,
  M = 5,
  N_T = round(N/2),
  min_acceptance_rate = 0,
  alpha = 0.9,
  marginal_kernel = FALSE,
  epsilon_min = 0,
  max_steps = 20,
  n_cores = (detectCores() - 2),
  upper_bound = NULL,
  lower_bound = NULL,
  output_dir = NULL,
  sample_tree = FALSE,
  M_sim = 1,
  min_distance_per_sim = TRUE,
  reobserve_resampled = FALSE,
  particle_collection_dir = NULL,
  keep_closest_tree_only = TRUE,
  allow_increase_in_eps = FALSE,
  delta_epsilon_wait = 5,
  delta_epsilon_min = 0,
  simulation_manipulator = NULL,
  tree_manipulator = NULL,
  previous_state = NULL,
  n_cores_dist = n_cores,
  ...
)
}
\arguments{
\item{const_params}{A set of constant parameters for the function call of 'new_simulation' (see its manual for details: new_simulation?)}

\item{p_gen}{Particle generator function. These provide additional parameters for the function call of 'new_simulation' (this would be the 'param.matrix' and 'seed').}

\item{s_gen}{Sample generator function. This returns a sampling schema for a given simulation (first argument).}

\item{target_y}{The target tree (object of class phylo.).}

\item{rho}{}

\item{N}{}

\item{M}{}

\item{N_T}{}

\item{min_acceptance_rate}{}

\item{alpha}{}

\item{marginal_kernel}{}

\item{epsilon_min}{}

\item{max_steps}{}

\item{n_cores}{}

\item{upper_bound}{}

\item{lower_bound}{}

\item{output_dir}{}

\item{sample_tree}{}

\item{M_sim}{}

\item{min_distance_per_sim}{}

\item{reobserve_resampled}{}

\item{particle_collection_dir}{}

\item{...}{}

\item{epsilon}{}
}
\value{

}
\description{
SMC-ABC on a phylogenetic tree
}
